Extract Property reader from the XML Reader into a separate class
